+++ src/types/plant.ts
@@ -1,0 +1,5 @@
+export interface Plant {
+    name: string;
+    nameEn?: string;
+    imgUrl?: string;
+}
